Drop service wrappers
These service wrappers only replicate the default systemd behaviour.
Running it directly simplifies it greatly because it also drops the need
to manage the rundir, which systemd will now take care of for us.

With the merge of
https://github.com/OpenVoxProject/trapperkeeper/pull/23, we can switch
to notify. Then the service will notify systemd when the service is
actually started.

Signed-off-by: Tim Meusel <tim@bastelfreak.de>