packaging: Switch from Perforce to OpenVoxProject releases
The Gemfile is a template. For every ezbake based build, the Gemfile is
copied into the build dir. Aftwards the gems are installed. the
packaging gem is used to publish packages, update repos and render
different templates, e.g. the systemd unit files for openvox-server and
openvox-db.

As far as I know, the erb template rendering is the only feature from
packaging that we use. I didn't want to deal with a new gem name yet,
the packaging rubygems namespace is already owned by Perforce: https://rubygems.org/gems/packaging

I assume in the long term we will get rid of packaging, so I didn't
bother with renaming the gem. Because of that, we cannot publish it to
rubygems.org right. Because of that the Gemfile now uses `github:` to
fetch the release from there.