introduce java_bin ezbake option
In
https://github.com/OpenVoxProject/ezbake/pull/2/commits/6b3ad939736bb8d0f0a22a715f398a49a678a8d3
we introduced a new configuration option for java_bin. To get this
working, we need to add it to ezbake itself. Otherwise the variable
won't exist and we cannot use it in templates:

```
ExecStart= $JAVA_ARGS $LOG_APPENDER \
  '-XX:OnOutOfMemoryError=kill -9 %p' -XX:+CrashOnOutOfMemoryError \
  -XX:ErrorFile="$LOGS_DIRECTORY/puppetserver_err_pid%p.log" \
  -cp "$INSTALL_DIR/puppet-server-release.jar" \
  clojure.main \
  -m puppetlabs.trapperkeeper.main \
  --config "$CONFIG" \
  --bootstrap-config "$BOOTSTRAP_CONFIG" \
  $TK_ARGS
```

(see the whitespace after `ExecStart=` where the java path should be.