Patch java_bin if overridden
In 9768a77de95ff19f66f4b328fc60efe69f5f3a6b the java binary was
overridden, but this didn't have a point because it's not available when
the files are rendered. They only have access to the ezbake config.

This takes an approach of patching the files after the fact. It writes
out backup files that are restored after. That is to deal with the fact
that our build process isn't clean and is reused between multiple
builds.

Fixes: 9768a77de95f ("Use virtual Java packages on Red Hat and set java_bin")
Signed-off-by: Tim Meusel <tim@bastelfreak.de>