ezbake: Introduce service-port variable
This adds a bit of a startup delay to ensure that systemd not just
starts the process and thinks the service is already up and running.

Previously this was handled by the wrapper script and Type=forking. This
commit fixes a regression from https://github.com/OpenVoxProject/ezbake/commit/b2de7c7a2e7704491a08665a9f53ded8009aa109

This also requires changes to the projects to provide the actual port.

* https://github.com/OpenVoxProject/openvox-server/pull/80
* https://github.com/OpenVoxProject/openvoxdb/pull/69

This is just a workaround and not a perfect solution. For systemd-based
systemd we should implement sd_notify and switch to Type=notify-reload.

https://www.freedesktop.org/software/systemd/man/latest/systemd.service.html#Type=